fix: ipywidgets 8 support
Ipywidgets 8 requires es6 class notation for models and views.
This required an upgrade of babel.

The js_helpers serializer was removed, because it was not base64
encoded. It only resulted in an error on ipywidgets 8 because of
the new echo functionality: https://ipywidgets.readthedocs.io/en/latest/changelog.html?highlight=echo#collaboration